+++ src/mocks/FilesHandler.js
@@ -13,6 +13,7 @@ const path = require("path");
 const Boom = require("boom");
 const requireAll = require("require-all");
 const watch = require("node-watch");
+const fsExtra = require("fs-extra");
 
 const { map, debounce } = require("lodash");
 
@@ -69,9 +70,7 @@ class FilesHandler {
 
   _resolveFolder(folder) {
     if (!folder) {
-      tracer.error(
-        'Please provide a path to a folder containing behaviors using the "behaviors" option'
-      );
+      tracer.error('Please provide a path to the folder containing mocks using the "path" option');
       throw Boom.badData("Invalid mocks folder");
     }
     if (path.isAbsolute(folder)) {
@@ -80,8 +79,13 @@ class FilesHandler {
     return path.resolve(process.cwd(), folder);
   }
 
+  _ensureFolder(folder) {
+    fsExtra.ensureDirSync(folder);
+    return folder;
+  }
+
   _loadFiles() {
-    this._path = this._resolveFolder(this._settings.get("behaviors"));
+    this._path = this._ensureFolder(this._resolveFolder(this._settings.get("path")));
     tracer.info(`Loading mocks from folder ${this._path}`);
     this._cleanRequireCacheFolder();
     this._files = requireAll({
@@ -110,7 +114,7 @@ class FilesHandler {
   }
 
   _onChangeSettings(changeDetails) {
-    if (changeDetails.hasOwnProperty("behaviors")) {
+    if (changeDetails.hasOwnProperty("path")) {
       this._loadFiles();
       this._switchWatch();
     } else if (changeDetails.hasOwnProperty("watch")) {